requirements.txt
torch
transformers
Pillow
opencv-python
einops
timm
